Added system strain
Attempted to add system strain as an attribute. It seems to work in terms of referencing, but I haven't checked automatically manipulating it, so I might be missing some logic somewhere - just use hp as a base if you need to add more.

Did some sizing for the custom resources, and added a couple.