Skill roll + ability mod
Adding ability mod to skill roll by making a new skill roll function that takes the name of the old one, and then calls the old one when it has determined the right ability modifier to use via dialogue to user, where they choose the attribute they want to use for a check.

Also added automatic fast forward for the skill roll.