Inventory Positioning
I was trying to do some complex stuff to optimize the positioning of the inventory and spell buttons. But it's just simpler to put the old button (equipped or attune button) and then make it vanish and not do anything. There's a little more space between the buttons than I'd like, but it's relatively minor.